+++ src/lerobot/scripts/server/helpers.py
@@ -110,7 +110,7 @@ def raw_observation_to_observation(
 
 def prepare_image(image: torch.Tensor) -> torch.Tensor:
     """Minimal preprocessing to turn int8 images to float32 in [0, 1], and create a memory-contiguous tensor"""
-    image = image.type(torch.float32) / 255
+    image = image.type(torch.bfloat16) / 255
     image = image.contiguous()
 
     return image
